requirements.txt
flet
numpy
scikit-learn
pillow
joblib
opencv-python-headless
